requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
blis==1.2.0
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
colorama==0.4.6
confection==0.1.5
cryptography==44.0.2
cymem==2.0.11
dj-database-url==2.0.0
Django==5.1.6
django-cors-headers==4.3.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docx2txt==0.8
drf-yasg==1.21.7
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
filelock==3.17.0
fsspec==2025.3.0
gunicorn==21.2.0
huggingface-hub==0.29.2
idna==3.10
inflection==0.5.1
Jinja2==3.1.5
joblib==1.4.2
langcodes==3.5.0
language_data==1.3.0
lxml==5.3.1
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.12
networkx==3.4.2
numpy==2.2.3
packaging==24.2
pandas==2.2.3
patool==3.1.3
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==11.1.0
preshed==3.0.9
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
PyPDF2==3.0.1
pypdfium2==4.30.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.0
pytz==2025.1
PyYAML==6.0.2
rarfile==4.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==3.4.1
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.3
srsly==2.5.1
sympy==1.13.1
thinc==8.3.4
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
wasabi==1.1.3
weasel==0.4.1
whitenoise==6.6.0
wrapt==1.17.2
openai==1.65.1
setuptools==69.0.3